Rename `main.py` to `cli.py`
There will be a `main.py` file in the future, but it will be used for
starting the GUI. Therefore, the current `main.py` was changed to
`cli.py`, as `cli.py` will be the entry point for the CLI.

Also removed the `main` function inside `main.py`, since we are not
currently using it. It may return in the near future, when we switch
from `argparse` to `click`.